+++ cypress/apiRequests/ShippingMethod.js
@@ -1,12 +1,18 @@
-export function createShippingRate(name, shippingZone) {
+import { getValueWithDefault } from "./utils/Utils";
+
+export function createShippingRate({ name, shippingZoneId }) {
   const mutation = `mutation{
     shippingPriceCreate(input:{
       name: "${name}"
-      shippingZone: "${shippingZone}"
+      shippingZone: "${shippingZoneId}"
       type: PRICE
     }){
       shippingMethod{
         id
+      }
+      errors{
+        field
+        message
       }
     }
   }`;
@@ -14,15 +20,23 @@ export function createShippingRate(name, shippingZone) {
 }
 
 export function createShippingZone(name, country, channelId) {
+  const channelsLines = getValueWithDefault(
+    channelId,
+    `addChannels:["${channelId}"]`
+  );
   const mutation = `mutation{
     shippingZoneCreate(input:{
       name: "${name}"
       countries: "${country}"
-      addChannels:["${channelId}"]
+      ${channelsLines}
     }){
       shippingZone{
         id
         name
+      }
+      errors{
+        field
+        message
       }
     }
   }`;
@@ -43,18 +57,24 @@ export function addChannelToShippingZone(shippingZoneId, channelId) {
   }`;
   return cy.sendRequestWithQuery(mutation);
 }
-export function addChannelToShippingMethod(shippingRateId, channelId, price) {
+export function addChannelToShippingMethod(
+  shippingRateId,
+  channelId,
+  price,
+  minProductPrice = 0
+) {
   const mutation = `mutation{
     shippingMethodChannelListingUpdate(id:"${shippingRateId}", input:{
       addChannels: {
         channelId:"${channelId}"
         price: ${price}
+        minimumOrderPrice:${minProductPrice}
       }
     }){
       shippingMethod{
         id
       }
-      shippingErrors{
+      errors{
         code
         message
       }
@@ -77,17 +97,33 @@ export function deleteShippingZone(shippingZoneId) {
 
 export function getShippingZones() {
   const query = `query{
-          shippingZones(first:100){
-            edges{
-              node{
-                name
-                id
-              }
-            }
-          }
+    shippingZones(first:100){
+      edges{
+        node{
+          name
+          id
         }
-        `;
+      }
+    }
+  }
+  `;
   return cy
     .sendRequestWithQuery(query)
     .then(resp => resp.body.data.shippingZones.edges);
+}
+
+export function getShippingZone(shippingZoneId) {
+  const query = `query{
+    shippingZone(id:"${shippingZoneId}"){
+      id
+      name
+      channels{
+        name
+        id
+      }
+    }
+  } `;
+  return cy
+    .sendRequestWithQuery(query)
+    .then(resp => resp.body.data.shippingZone);
 }