+++ cypress/elements/channels/add-channel-form-selectors.js
@@ -8,5 +8,7 @@ export const ADD_CHANNEL_FORM_SELECTORS = {
   currencyValidationMessage: "[data-testid='currency-text-input-helper-text']",
   slugValidationMessage: "[data-testid='slug-text-input-helper-text']",
   currencyAutocompleteDropdown:
-    "[data-test='singleautocomplete-select-option'][data-test-type='custom']"
+    "[data-test='singleautocomplete-select-option'][data-test-type='custom']",
+  addShippingZoneButton: '[data-test-id="add-shipping-zone-button"]',
+  shippingAutocompleteSelect: "[data-test-id='shippingAutoCompleteSelect']"
 };