+++ cypress/integration/allEnv/orders/orders.js
@@ -99,6 +99,8 @@ describe("Orders", () => {
   });
 
   it("should create order with selected channel", () => {
+    // Remove login as admin after fixing SALEOR-3154
+    cy.clearSessionData().loginUserViaRequest();
     cy.visit(urlList.orders)
       .get(ORDERS_SELECTORS.createOrder)
       .click();