+++ cypress/utils/products/productsUtils.js
@@ -1,6 +1,11 @@
 import * as attributeRequest from "../../apiRequests/Attribute";
 import * as categoryRequest from "../../apiRequests/Category";
 import * as productRequest from "../../apiRequests/Product";
+import {
+  createTypeProduct,
+  deleteProductType,
+  getProductTypes
+} from "../../apiRequests/productType";
 import { deleteAttributesStartsWith } from "../attributes.js/attributeUtils";
 
 export function createProductInChannel({
@@ -69,7 +74,7 @@ export function createTypeAttributeAndCategoryForProduct(
     .createAttribute(name, attributeValues)
     .then(attributeResp => {
       attribute = attributeResp;
-      productRequest.createTypeProduct({ name, attributeId: attributeResp.id });
+      createTypeProduct({ name, attributeId: attributeResp.id });
     })
     .then(productTypeResp => {
       productType = productTypeResp;
@@ -81,12 +86,13 @@ export function createTypeAttributeAndCategoryForProduct(
     });
 }
 export function deleteProductsStartsWith(startsWith) {
+  deleteAttributesStartsWith(startsWith);
+  cy.deleteElementsStartsWith(deleteProductType, getProductTypes, startsWith);
   cy.deleteElementsStartsWith(
-    productRequest.deleteProductType,
-    productRequest.getProductTypes,
+    attributeRequest.deleteAttribute,
+    attributeRequest.getAttributes,
     startsWith
   );
-  deleteAttributesStartsWith(startsWith);
   cy.deleteElementsStartsWith(
     categoryRequest.deleteCategory,
     categoryRequest.getCategories,